+++ Java/tollCalculator/src/main/java/com/acme/tollCalculator/TollCalculatorController.java
@@ -5,12 +5,9 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.acme.tollCalculator.dataTypes.Vehicle;
 import com.acme.tollCalculator.dataTypes.VehicleDeserializer;
 
 import java.util.*;
-//import java.util.concurrent.*;
-
 @RestController
 public class TollCalculatorController {
 
@@ -33,26 +30,5 @@ public class TollCalculatorController {
         var v = vehicleDeserializer.getVehicleFromString(vehicle);
         return tollCalculator.getTollFee(v, dates);
     }
-    
-	@GetMapping("/")
-	public String index() {
-		return "Greetings and hello from Spring Boot!";
-	}
-
-    @GetMapping("/vehicle-types")
-    public String[] getVehicleTypes() {
-        return vehicleDeserializer.getTypeNames();
-    }
-
-    @GetMapping("/current-date")
-    public String getCurrentDate() {
-        return new Date().toString();
-    }
-
-    @GetMapping("/parse-date")
-    public String parseDate(@RequestParam @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME) Date date) {
-        return date.toString();
-    }
-    
 
 }
